Add files via upload
Remove the 4th icon at the bottom of the page of Google as we already have a link to email and other than that the on when we log into the phone right with three icons look good. When we have 4th the 4th one goes at the next line. So it is like 3 on the first line and the 4th one on the next line. So it doesn't look good. So I took off Google completely the icon of Google completely. So what will happen now is we will have only three icons, 1 of Github, one of linkedin and 1 of Twitter