Add files via upload
Created all the three separate blog 1 blog 2 blog 3 respectively, added all the comments that are actually related to the blog. I read the actual names of the commenters, therefore. their photos created a new folder in the image folder. Also in the blog section on the right side we have an archive section worked on that. That's it. The main changes are still left